% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDist.r
\name{CalcDist}
\alias{CalcDist}
\title{Distance calculation function from Norman Abrahamson's HAZ program.}
\usage{
CalcDist(faultW, faultLen, seismoDepth = 0, ft.fltGrid_x, ft.fltGrid_y,
  ft.fltGrid_z, ft.fltGrid_x1, ft.fltGrid_y1, ft.fltGrid_z1, ft.fltGrid_x2,
  ft.fltGrid_y2, ft.fltGrid_z2, ft.fltGrid_x3, ft.fltGrid_y3, ft.fltGrid_z3,
  ft.fltGrid_x4, ft.fltGrid_y4, ft.fltGrid_z4, ft.fltGrid_Rrup, ft.fltGrid_Rjb,
  ystep = 1)
}
\arguments{
\item{faultW}{Fault width}

\item{faultLen}{Fault length.}

\item{seismoDepth}{seismogenic depth (fault minDepth in fault file).}

\item{ft.fltGrid_x}{Fault grid coordinate x.}

\item{ft.fltGrid_y}{Fault grid coordinate y.}

\item{ft.fltGrid_z}{Fault grid coordinate z.}

\item{ft.fltGrid_x1}{Fault grid coordinate x1.}

\item{ft.fltGrid_y1}{Fault grid coordinate y1.}

\item{ft.fltGrid_z1}{Fault grid coordinate z1.}

\item{ft.fltGrid_x2}{Fault grid coordinate x2.}

\item{ft.fltGrid_y2}{Fault grid coordinate y2.}

\item{ft.fltGrid_z2}{Fault grid coordinate z2.}

\item{ft.fltGrid_x3}{Fault grid coordinate x3.}

\item{ft.fltGrid_y3}{Fault grid coordinate y3.}

\item{ft.fltGrid_z3}{Fault grid coordinate z3.}

\item{ft.fltGrid_x4}{Fault grid coordinate x4.}

\item{ft.fltGrid_y4}{Fault grid coordinate y4.}

\item{ft.fltGrid_z4}{Fault grid coordinate z4.}

\item{ft.fltGrid_Rrup}{Fault grid rupture distance.}

\item{ft.fltGrid_Rjb}{Fault grid JB distance.}

\item{ystep}{ystep in fault input file.}
}
\value{
A list will be returned, includingHWFlag, hypoDepth, ZTOR, RupWidth, RupLen, seismoDepth,
distJB, distRup, distSeism, distepi, disthypo, Rx, Ry, Ry0.
}
\description{
\code{CalcDist} computes the distance parameters for the GMPEs.
                (e.g. Rjb, Rrup, ZTOR, Rx, Ry, Ry0)
}
\examples{
CalcDist(14.14, 8.72, seismoDepth=0, array(0, c(4,4)), array(0, c(4,4)), array(0, c(4,4)),
array(1, c(4,4)), array(1, c(4,4)), array(1, c(4,4)),
array(2, c(4,4)), array(2, c(4,4)), array(2, c(4,4)),
array(3, c(4,4)), array(3, c(4,4)), array(3, c(4,4)),
array(4, c(4,4)), array(4, c(4,4)), array(4, c(4,4)),
array(2, c(4,4)), array(2, c(4,4)), ystep=1.0)


}
