% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertCoordinates2.r
\name{ConvertCoordinates2}
\alias{ConvertCoordinates2}
\title{Function for converting latitude/longitude to x/y coordinates on the basis of a spherical earth
from Norman Abrahamson's HAZ program.}
\usage{
ConvertCoordinates2(siteX, siteY, ft.Z, ft.Lat, ft.Long, nDD, pt.line)
}
\arguments{
\item{siteX}{Site coordinate X (longitude).}

\item{siteY}{Site coordinate Y (latitude).}

\item{ft.Z}{Fault line coordinate (top of fault).}

\item{ft.Lat}{Fault line coordinates (latitude).}

\item{ft.Long}{Fault line coordinate (longitude).}

\item{nDD}{number of points on down dip direction.}

\item{pt.line}{number of points on fault line.}
}
\value{
A list will be return, including siteX, siteY, xFlt(,), yFlt(,), zFlt(,), x0, y0, z0.
}
\description{
\code{ConvertCoordinates2} returns fault coordinates in grid(km).
}
\examples{
ConvertCoordinates2(121.5, 25, array(c(0,0),c(1,1,2)),
                    array(c(24.74554,24.7901),c(1,1,2)),
                    array(c(120.9609, 121.0747),c(1,1,2)), 2, 2)
ConvertCoordinates2(121.0, 24.5, array(c(0,0),c(1,1,2)),
                    array(c(24.74554,24.7901),c(1,1,2)),
                    array(c(120.9609, 121.0747),c(1,1,2)), 2, 2)

}
