% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{FT}
\alias{FT}
\title{Fault spatial data frame.}
\format{A spatial data frame of Shanchiao fault :
\describe{
  \item{FAULT_NAME}{fault name}
  \item{FAULT_TYPE}{fault type}
  \item{LENGTH_KM}{fault length}
  \item{FAULT_DIP1}{fault dip}
  \item{DEPTH1_KM}{fault depth}
  \item{WIDTH_KM}{fault width}
  ...
}}
\usage{
FT
}
\description{
A dataset containing the Shanchiao fault.
}
\keyword{datasets}
