% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfaultcoord.r
\name{getfaultcoord}
\alias{getfaultcoord}
\title{Function for get fault coordinates from spatial data.frame.}
\usage{
getfaultcoord(faultlines, id, c.id, c.fault.top, c.fault.thick, c.dip)
}
\arguments{
\item{faultlines}{a spatial data frame of fault.}

\item{id}{ID from fault.}

\item{c.id}{Column name of the fault id.}

\item{c.fault.top}{Column name of depth of top of fault(km).}

\item{c.fault.thick}{Column name of the fault thick(km) (seismogenic thickness).}

\item{c.dip}{Column name of the fault dip(degree).}
}
\value{
A list will be return, including dip, faultwidth, fZ(,), fLat(,), fLong(,), nDD, pt.line.
}
\description{
\code{getfaultcoord} returns coordinates of fault from spatial data frame.
}
\examples{
getfaultcoord(FT, 1, "ID", "DEPTH1_KM", "WIDTH_KM", "FAULT_DIP1")

}
