% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfaultdist.r
\name{getfaultdist}
\alias{getfaultdist}
\title{Function for get fault distances.}
\usage{
getfaultdist(fault, site, seismoDepth = 0)
}
\arguments{
\item{fault}{fault coordinates with list name of fZ, fLat, fLong, nDD, pt.line.}

\item{site}{data.frame with x and y.}

\item{seismoDepth}{seismogenic depth (fault minDepth in fault file).}
}
\value{
A data.frame will be return, including column of HWFlag, hypoDepth, ZTOR, RupWidth, RupLen,
seismoDepth, distJB, distRup, distSeism, distepi, disthypo, Rx, Ry, Ry0.
}
\description{
\code{getfaultdist} returns distances of the fault.
}
\examples{
site <- data.frame(x=121.001, y=24.7794)
fault <- getfaultcoord(FT, 1, "ID", "WIDTH_KM", "FAULT_DIP1")
getfaultdist(fault, site, seismoDepth=0)


}
