% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAZDist.r
\docType{package}
\name{HAZDist}
\alias{HAZDist}
\alias{.onUnload}
\alias{HAZDist-package}
\title{HAZDist: Distance calculation funciton using Norman Abrahamson's HAZ fortran code.}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{library path}
}
\description{
The HAZDist package provides distance calculation functions:
SetFltBottom, ConvertCoordinates2, calcfltgrid, CalcDist, getfaultcoord, getfaultdist
}
