% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFltBottom.r
\name{SetFltBottom}
\alias{SetFltBottom}
\title{Function for setting bottom of fault for standard faults (source type 1)
from Norman Abrahamson's HAZ program.}
\usage{
SetFltBottom(dip, faultwidth, ft.Z, ft.Lat, ft.Long, pt.line)
}
\arguments{
\item{dip}{Dip angle of the fault plane.}

\item{faultwidth}{Faultwidth(km) = faultThick1 = seismogenic thickness (fault widths in fault file).}

\item{ft.Z}{Fault line coordinate (top of fault).}

\item{ft.Lat}{Fault line coordinates (latitude).}

\item{ft.Long}{Fault line coordinate (longitude).}

\item{pt.line}{Number of points of the fault line.}
}
\value{
A list will be return, including dip, faultwidth, fZ(,), fLat(,), fLong(,), nDD.
}
\description{
\code{SetFltBottom} returns coordinates on bottom of fault.
}
\examples{
SetFltBottom(45, 15, c(0,0), c(24.74554,24.7901), c(120.9609, 121.0747), 2)
SetFltBottom(30, 15, c(0,0), c(24.74554,24.7901), c(120.9609, 121.0747), 2)


}
